Refactor: Migrate Login and Register to 'auth' folder and remove modal.css in favor of TailwindCSS
