+++ src/components/Modal.tsx
@@ -1,6 +1,4 @@
 import React, { JSX, useEffect, useRef } from 'react';
-import '../css/modal/modal.css'
-
 
 export function Modal({ children, onClose, modalTitle, style, modalId }: { children: React.ReactNode; onClose: () => void; modalTitle?: string; style?: React.CSSProperties; modalId?: string }): JSX.Element {
     const modalRef = useRef<HTMLDivElement>(null);
@@ -22,8 +20,8 @@ export function Modal({ children, onClose, modalTitle, style, modalId }: { child
     }, []);
 
     return <>
-        <div className="modal-fade" data-modal-id={modalId}>
-            <div className="modal-content" ref={modalRef} onClick={(e) => e.stopPropagation()} style={style}>
+        <div className="fixed top-0 left-0 w-full h-full bg-[rgba(0, 0, 0, .8)] flex items-center justify-center z-10" data-modal-id={modalId}>
+            <div className="bg-[#454545] p-5 max-w-7xl rounded-2xl" ref={modalRef} onClick={(e) => e.stopPropagation()} style={style}>
                 <div className=" p-4 text-center">
                     <h3 className="text-xll font-semibold">{modalTitle}</h3>
                 </div>
