--- src/css/modal/modal.css
@@ -1,19 +1,0 @@
-.modal-fade {
-    position: fixed;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 100%;
-    background: rgba(0, 0, 0, .8);
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    z-index: 9;
-}
-
-.modal-fade .modal-content {
-    background: #454545;
-    padding: 20px;
-    max-width: 1200px;
-    border-radius: 10px;
-}
